api/requirements.txt
flask
flask-smorest
python-dotenv
sqlalchemy
flask-sqlalchemy
flask-marshmallow
flask-jwt-extended
passlib